moved the functions out of the app.jsx and into its own util file
