+++ src/App.jsx
@@ -11,56 +11,27 @@ import {
 import useImage from 'use-image';
 import maps from './maps.json';
 
-import soundfile from './sounds/huntingMusic.mp3';
 import { QuizSection } from './quiz/main';
+import {
+  controlAudio,
+  generateGiveUpMessage,
+  renderLoadingScreen,
+  resetTriggers,
+  triggerRoomUnlock
+} from './util';
 
 // *****************************************************
 
 // Turn off HUNT_MODE to enable tools to get the x/y/radius of the eggs
 const HUNT_MODE = true;
-
-// *****************************************************
-
-const setImageForRoom = (roomToChange, imageToChangeTo) => {
-  let image = new window.Image();
-  image.src = imageToChangeTo;
-  roomToChange.image = image;
-};
-
-// Preload all the maps, also how the images are generated
-Object.keys(maps).forEach(key => {
-  // Fake images for populating the inspect element with junk
-  setImageForRoom(maps[key], maps[key].fakeImageName)
-  setImageForRoom(maps[key], maps[key].imageName)
-});
-
-// *****************************************************
-
 const height = window.innerHeight;
 const width = window.innerWidth;
-
-// AUDIO SETTINGS **************************
-
-const audioElement = new Audio(soundfile);
-
-const controlAudio = (command) => {
-  if (command === 'play'){
-    audioElement.loop = true;
-    audioElement.volume = 0.2;
-    audioElement.play();
-  } else if (command === 'stop'){
-    audioElement.pause();
-    audioElement.currentTime = 0;
-  }
-};
-
-// ******************************************
 
 export function App() {
   const [status, setStatus] = useState('loading');
 
   const [name, setName] = useState('');
-  const [score, setScore] = useState(50);
+  const [score, setScore] = useState(0);
   const [maxScore, setMaxScore] = useState(50);
   const [foundEggs, setFoundEggs] = useState([]);
   const [foundKeys, setFoundKeys] = useState([]);
@@ -82,51 +53,7 @@ export function App() {
   const [checkmark] = useImage('checkmark.gif');
   const [congratulations] = useImage('Congratulations.png');
 
-  //************************************************************
-  // THIS IS THE CODE TO CHANGE THE JPEG FILE FOR ROOMS WITH CLOSED THEN OPEN DOORS
-  const triggerRoomUnlock = (roomWhereKeyIsFound) => {
-    switch(roomWhereKeyIsFound) {
-      case 'BEDROOMCLOSET':
-        alert('You found the key that unlocks the door in the living room! Go check it out!');
-        setImageForRoom(maps.LIVINGROOM, maps.IMAGECHANGES.livingRoomUnlockedImage);
-        maps.LIVINGROOM.up = 'LIVINGROOMCLOSET';
-        break;
-      case 'KITCHENCORNER':
-        alert('You found the key that unlocks the door in the bedroom! Go check it out!');
-        setImageForRoom(maps.BEDROOMCORNER, maps.IMAGECHANGES.bedroomCornerUnlockedImage);
-        maps.BEDROOMCORNER.up = 'BEDROOMCLOSET';
-        break;
-      case 'MYSTERY':
-        maps.KITCHENCORNER.mystery = 'MYDOOR';
-        break;
-      default:
-        alert('You have found the test trigger. This alert is all that happens because of it. You really shouldn\'t be seeing this.');
-    }
-  };
-
-// revert all the changes that could be made in triggerRoomUnlock
-  const resetTriggers = () => {
-    setImageForRoom(maps.LIVINGROOM, maps.IMAGECHANGES.livingRoomLockedImage);
-    setImageForRoom(maps.BEDROOMCORNER, maps.IMAGECHANGES.bedroomCornerLockedImage);
-
-    if (maxScore === 250) {
-      setMaxScore(50);
-    }
-
-    if (maps.LIVINGROOM.up) {
-      delete maps.LIVINGROOM.up;
-    }
-    if (maps.BEDROOMCORNER.up) {
-      delete maps.BEDROOMCORNER.up;
-    }
-    if (maps.KITCHENCORNER.mystery) {
-      delete maps.KITCHENCORNER.mystery;
-    }
-  };
-
-  // ******************************************************************
-
-  let handleImageDrag = event => {
+  const handleImageDrag = event => {
     setEggX((event.target.attrs.x- imageX) / scale );
     setEggY((event.target.attrs.y  / scale));
   };
@@ -168,34 +95,9 @@ export function App() {
     }
   }, [image]);
 
-  let changeLocation = (newLocationName) => {
+  const changeLocation = (newLocationName) => {
     setCurrentLocation(maps[newLocationName])
   };
-
-  const generateGiveUpMessage = () => {
-    if (score === 50) {
-      alert(`${name} is a super hunter who found all 50 eggs!\nWOW!! Thanks for playing, and hope to see you again, soon!`);
-    } else {
-      alert(`${name} found ${score} out of 50 eggs!\nThanks for playing!`);
-    }
-  };
-
-  function renderLoadingScreen() {
-    let screen = [];
-    for (let i = 0; i < 13; i++) {
-      screen.push(<div
-        style={{
-          position: "absolute",
-          top: `${Math.floor((Math.random() * 1000))}px`,
-          left: `${Math.floor((Math.random() * 1000))}px`,
-          transform: `rotate(${Math.floor((Math.random() * 360))}deg) 
-          scale(${Math.floor((Math.random() * 10))}, ${Math.floor((Math.random() * 10))}`,
-        }}>
-        LOADING
-      </div>);
-    }
-    return screen;
-  }
 
   if (status === 'loading') {
     return (<div>
@@ -276,13 +178,13 @@ export function App() {
           }}
           onClick={() => {
             controlAudio('stop');
-            generateGiveUpMessage();
+            generateGiveUpMessage(score, name);
             setStatus('landing');
             setName('');
             setScore(0);
             setFoundEggs([]);
             setFoundKeys([]);
-            resetTriggers();
+            resetTriggers(maxScore, setMaxScore);
             setCurrentLocation(maps.LIVINGROOM);
           }
         }
@@ -413,7 +315,7 @@ export function App() {
                     y={ (egg.eggY * scale)}
                     radius={ egg.eggRadius * scale}
                     onClick={() => {
-                      setScore(score+1)
+                      setScore(score+1);
                       setFoundEggs([`${currentLocation.name}egg${i}`, ...foundEggs])
                     }}
                     key={`${currentLocation.name}egg${i}`}
@@ -455,7 +357,7 @@ export function App() {
                       y={(key.keyY * scale)}
                       radius={key.keyRadius * scale}
                       onClick={() => {
-                        triggerRoomUnlock(currentLocation.name)
+                        triggerRoomUnlock(currentLocation.name);
                         setFoundKeys([`${currentLocation.name}key${i}`, ...foundKeys])
                       }}
                       key={`${currentLocation.name}key${i}`}
@@ -480,7 +382,7 @@ export function App() {
               scaleX={0.5}
               scaleY={0.5}
               onClick={() => {
-                triggerRoomUnlock('MYSTERY')
+                triggerRoomUnlock('MYSTERY');
                 setMaxScore(250)
               }}
             /> : null
